APR 15
- Initialize Firebase in the main application entry point.
- Update `GroupItem` and `Item` models to handle weight as a `double` and fix cost parsing.
- Refactor `ItemScreen` to include category-based filtering and a real-time list view for items using Firestore.
- Improve item creation in `addItem.dart` by adding field clearing, description sanitization, and better state management for unit selection.
- Update `UnitSelector` to manage its own unit state internally.
- Enhance UI scrolling behavior and loading indicators across the application.

+++ lib/models/groupItem.dart
@@ -4,10 +4,10 @@ class GroupItem {
   List<Item>? items;
   String? description;
   int? cost;
-  int? weight;
+  double? weight;
 
   GroupItem(this.items, this.description, this.cost, this.weight) {
-    int accumulatedWeight = 0;
+    double accumulatedWeight = 0;
 
     items?.forEach((e) {
       if (e.weight != null){
